% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTest.r
\name{hdcrpt}
\alias{hdcrpt}
\title{High dimensional censored rank partial test}
\usage{
hdcrpt(
  x,
  y,
  status,
  z,
  smooth = c("sigmoid", "gaussian"),
  h = NULL,
  covariate_dependence = TRUE
)
}
\arguments{
\item{x}{The design matrix.}

\item{y}{The survival outcome.}

\item{status}{The right censoring indicator.}

\item{z}{The control vector, which is the control factors multiplied by the estimated coefficients.}

\item{smooth}{The smooth function used. "sigmoid" represents the sigmoid prime kernel function and "gaussian" represents the gaussian kernel function.}

\item{h}{The bandwidth of the control vector. The default value is the standard deviation of z divided by the square root of sample size n.}

\item{covariate_dependence}{An indicator of whether the censoring mechanism is dependent on covariates.}
}
\value{
A list.
\itemize{
    \item ts - Test statistic.
    \item pval - p value.
}
}
\description{
High dimensional censored rank partial test
}
\examples{
set.seed(0)
n <- 150
p <- 550
x <- matrix(rnorm(n * p), n, p)
u <- matrix(rnorm(n * p), n, p)
alpha <- c(rep(1, 5), rep(0, p - 5))
beta <- c(rep(1, 5), rep(0, p - 5))
y <-  u \%*\% alpha + x \%*\% beta + rnorm(n)
status <- sample(c(0, 1), n, replace = TRUE, prob = c(0.2, 0.8))
fit <- sprfabs(y, u, status)
alphahat <- fit$opttheta
z <- u \%*\% alphahat
test_result_sigmoid <- hdcrpt(x, y, status, z, "sigmoid")
print(test_result_sigmoid)
test_result_gaussian <- hdcrpt(x, y, status, z, "gaussian")
print(test_result_gaussian)
}
\seealso{
\code{\link{hdcrt}}, \code{\link{hdcrt_dc}}
}
