% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTest.r
\name{hdcrt}
\alias{hdcrt}
\title{High dimensional censored rank test}
\usage{
hdcrt(x, y, status, covariate_dependence = TRUE)
}
\arguments{
\item{x}{The design matrix.}

\item{y}{The survival outcome.}

\item{status}{The right censoring indicator.}

\item{covariate_dependence}{An indicator of whether the censoring mechanism is dependent on covariates.}
}
\value{
A list.
\itemize{
    \item ts - Test statistic.
    \item pval - p value.
}
}
\description{
High dimensional censored rank test
}
\examples{
set.seed(0)
n <- 150
p <- 550
x <- matrix(rnorm(n * p), n, p)
beta <- c(rep(1, 5), rep(0, p - 5))
y <-  x \%*\% beta + rnorm(n)
status <- sample(c(0, 1), n, replace = TRUE, prob = c(0.2, 0.8))
test_result <- hdcrt(x, y, status)
print(test_result)
}
\seealso{
\code{\link{hdcrpt}}, \code{\link{hdcrt_dc}}
}
