% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprfabs.r
\name{sprfabs}
\alias{sprfabs}
\title{A forward and backward stagewise algorithm for high-dimensional spr problem.}
\usage{
sprfabs(
  y,
  x,
  status,
  eps = 0.01,
  xi = 1e-10,
  maxIter = 3000,
  sigma = NULL,
  weight = NULL,
  nmax = NULL,
  lam_m = NULL,
  message = TRUE
)
}
\arguments{
\item{y}{The response.}

\item{x}{The design matrix.}

\item{status}{The right censoring indicator.}

\item{eps}{The step size for updating coefficients. Default is \code{eps = 0.01}.}

\item{xi}{The threshold for qfabs. Default is \code{xi = 1e-10}.}

\item{maxIter}{The maximum number of outer-loop iterations allowed. Default is \code{maxIter = 3000}.}

\item{sigma}{The tuning parameter in the Sigmoid function. Default is \code{NULL}.}

\item{weight}{An optional weights. Default is 1 for each observation.}

\item{nmax}{Limit the maximum number of variables in the model. When exceed this limit, program will early stopped. Default is \code{NULL}.}

\item{lam_m}{The ratio of the minumum lambda and the maximum lambda. Default is \code{NULL}.}

\item{message}{An indicator of whether print warning messages.}
}
\value{
A list.
\itemize{
  \item theta - The estimation of covariates.
  \item opttheta - The optimal estimation of covariates.
  \item lambda - lambda sequence.
  \item direction - direction sequence.
  \item iter - Iterations.
  \item bic - The EBIC for each solution.
  \item loss - loss for each solution.
  \item df - Number of nonzero coefficients.
  \item opt - Position of the optimal lambda based on EBIC.
}
}
\description{
A forward and backward stagewise algorithm for high-dimensional spr problem.
}
\examples{
set.seed(0)
n <- 150
p <- 550
x <- matrix(rnorm(n * p), n, p)
alpha <- c(rep(1, 5), rep(0, p - 5))
y <-  x \%*\% alpha + rnorm(n)
status <- sample(c(0, 1), n, replace = TRUE, prob = c(0.2, 0.8))
fit <- sprfabs(y, x, status)
alphahat <- as.vector(fit$opttheta)
print(which(alphahat != 0))
print(alphahat[alphahat != 0])
}
