% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTest.r
\name{hdcrt_dc}
\alias{hdcrt_dc}
\title{High dimensional censored rank test with double censored}
\usage{
hdcrt_dc(x, y, status_left, status_right)
}
\arguments{
\item{x}{The design matrix.}

\item{y}{The survival outcome.}

\item{status_left}{The left censoring indicator.}

\item{status_right}{The right censoring indicator.}
}
\value{
A list.
\itemize{
    \item ts - Test statistic.
    \item pval - p value.
}
}
\description{
High dimensional censored rank test with double censored
}
\examples{
set.seed(0)
n <- 150
p <- 550
x <- matrix(rnorm(n * p), n, p)
beta <- c(rep(1, 5), rep(0, p - 5))
y <-  x \%*\% beta + rnorm(n)
status_left <- y >= qnorm(0.1) * sqrt(6)
status_right <- y <= qnorm(0.9) * sqrt(6)
test_result <- hdcrt_dc(x, y, status_left, status_right)
print(test_result)
}
\seealso{
\code{\link{hdcrpt}}, \code{\link{hdcrt}}
}
